Archive/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
pydantic==2.8.2
requests==2.32.3
pyzmq==26.0.3
gspread==6.1.2
google-auth==2.34.0
pandas==2.3.2
pandas-ta==0.4.71b0